+++ src/components/addCardSection.tsx
@@ -8,7 +8,7 @@ import {
 import { bank } from "../core/bank";
 import { AddCardForm } from "./addCardForm";
 import { BankMenu } from "./bankMenu";
-import { Card } from "./common/card";
+import { Card } from "./common/card/card";
 import { ModalBackDrop } from "./common/modal/modalBackDrop";
 import { ModalBox } from "./common/modal/modalBox";
 import { ModalContent } from "./common/modal/modalContent";
