+++ src/components/common/card/card.tsx
@@ -1,6 +1,6 @@
 import styled from "styled-components";
-import { CardColor, CardInfo } from "../../type/card";
-import { CardNumberIndex } from "../../type/input";
+import { CardColor, CardInfo } from "../../../type/card";
+import { CardNumberIndex } from "../../../type/input";
 
 export interface CardProps extends CardInfo {}
