+++ src/pages/cardListPage.tsx
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 import { Header } from "../components/common/Header";
 import { useState } from "react";
 import { getData } from "../utils/localStorage";
-import { Card, CardProps } from "../components/common/card";
+import { Card, CardProps } from "../components/common/card/card";
 import { ReactComponent as BackButtonIc } from "../assets/backButtonIc.svg";
 import { PAGE_HEADER } from "../constants";
 import { PATH } from "../constants/path";
