--- src/stories/NameInput.stories.tsx
@@ -1,19 +1,0 @@
-import React from "react";
-// import { Story } from "@storybook/react";
-
-// import { InputBox, InputBoxProps } from "../components/common/InputBox";
-
-// export default {
-//   component: InputBox,
-//   title: "Task",
-// };
-
-// const Template: Story<InputBoxProps> = (args: InputBoxProps) => (
-//   <InputBox {...args} />
-// );
-
-// export const Default = Template.bind({});
-// Default.args = {
-//   // labelText: "카드 소유자 이름(선택)",
-//   // placeholder: "카드에 표시된 이름과 동일하게 입력하세요.",
-// };
