+++ src/utils/localStorage.ts
@@ -1,4 +1,4 @@
-import { CardProps } from "../components/common/card";
+import { CardProps } from "../components/common/card/card";
 
 export function setData(data: CardProps) {
   const prevData = getData();
